test(data): move `BytesOp` etc. into test utils module
