+++ data/Cargo.toml
@@ -7,6 +7,9 @@ rust-version.workspace = true
 [lints]
 workspace = true
 
+[features]
+unstable-test-utils = ["dep:proptest"]
+
 [dependencies]
 bincode.workspace = true
 blake3.workspace = true
@@ -19,6 +22,10 @@ memmap2.workspace = true
 range-collections.workspace = true
 bytes.workspace = true
 unty.workspace = true
+
+[dependencies.proptest]
+workspace = true
+optional = true
 
 [dev-dependencies]
 proptest.workspace = true
