+++ data/src/components/bytes.rs
@@ -6,6 +6,8 @@
 //!
 //! See [`Bytes`] for more details.
 
+pub mod test_utils;
+
 use std::borrow::Borrow;
 use std::cell::Cell;
 use std::cell::RefCell;
