+++ data/src/components/vector/tests.rs
@@ -20,8 +20,8 @@ use crate::components::atom::tests::AtomMutOp;
 use crate::components::atom::tests::AtomOp;
 use crate::components::bytes::Bytes;
 use crate::components::bytes::BytesMode;
-use crate::components::bytes::tests::BytesMutOp;
-use crate::components::bytes::tests::BytesOp;
+use crate::components::bytes::test_utils::BytesMutOp;
+use crate::components::bytes::test_utils::BytesOp;
 use crate::components::vector::VectorMode;
 use crate::foldable::Foldable;
 use crate::foldable::Unfoldable;
